22594: Remove jenkins-image-k8s-tests
We never set up a Jenkins job to actually use this image, as evidenced
by the fact it never got updated past Debian 10. So much has changed
about Kubernetes in the meantime that I expect if we revisit support or
testing, we'll want to look at it with fresh eyes.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>