22594: Remove jenkins-image-arvados-formula-tests
This image has been used to test our Arvados Salt formula. This image
actually does use a tool for that that isn't included in other
images (kitchen). But we haven't actually run these formula tests in
three years despite making arvados-formula changes in the meantime—we
mostly lean on test-provision to be a big integration test for us.

We're not likely to revisit this since we plan to move to an
Ansible-based installer. Just remove this image.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>